DAO замен на JPA,удалены add-user и контроллеры связанные с ним ввидуненужность  осталось доделать обновление ролли в выпадающем списке

+++ src/main/java/ru/igor/PP_3_1_3/repositories/PeopleRepository.java
@@ -4,9 +4,11 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 import ru.igor.PP_3_1_3.model.User;
 
+import java.util.List;
 import java.util.Optional;
 
 @Repository
-public interface PeopleRepository extends JpaRepository<User, Integer> {
+public interface PeopleRepository extends JpaRepository<User, Long> {
     Optional<User> findByUsername(String username);
 }
+
